0.2.0 - Adding tuple support
