+++ mix.exs
@@ -4,7 +4,7 @@ defmodule PhpAssocMap.MixProject do
   def project do
     [
       app: :php_assoc_map,
-      version: "0.1.0",
+      version: "0.2.0",
       name: "PhpAssocMap",
       description: "Library that parses PHP's associative array into Elixir's map.",
       source_url: "https://github.com/nicklayb/php_assoc_map",
