+++ test/tuple_test.exs
@@ -1,23 +1,8 @@
-defmodule PhpAssocMapTest do
+defmodule TupleTest do
   use ExUnit.Case
   doctest PhpAssocMap
 
   @flatten_source "['lvl_1_1'=>['lvl_2_1'=>1,'lvl_2_2'=>'Single quoted string','lvl_2_3'=>\"Double quoted string\"],'lvl_1_2'=>false]"
-
-  test "flatten associative array" do
-    source = """
-    [
-      'lvl_1_1' => [
-        'lvl_2_1' => 1,
-        'lvl_2_2' => 'Single quoted string',
-        'lvl_2_3' => "Double quoted string"
-      ],
-      'lvl_1_2' => false
-    ]
-    """
-
-    assert @flatten_source == PhpAssocMap.flatten_assoc(source)
-  end
 
   test "parse unflatten array" do
     source = """
@@ -35,24 +20,26 @@ defmodule PhpAssocMapTest do
     ]
     """
 
-    expected = %{
-      "lvl_1_1": %{
-        "lvl_2_1": 1,
-        "lvl_2_2": "Single quoted string",
-        "lvl_2_3": "Double quoted string"
+    expected = [
+      {"lvl_1_1", [
+          {"lvl_2_1", 1},
+          {"lvl_2_2", "Single quoted string"},
+          {"lvl_2_3", "Double quoted string"},
+        ]
       },
-      "lvl_1_2": false,
-      "lvl_1_3": %{
-        "lvl_2_1": true,
-        "lvl_2_2": 54.12
+      {"lvl_1_2", false},
+      {"lvl_1_3", [
+          {"lvl_2_1", true},
+          {"lvl_2_2", 54.12},
+        ]
       },
-    }
+    ]
 
-    assert PhpAssocMap.to_map(source) == expected
+    assert PhpAssocMap.to_tuple(source) == expected
   end
 
   test "flatten a flatten associative array" do
-    assert @flatten_source == PhpAssocMap.flatten_assoc(@flatten_source)
+    assert @flatten_source == PhpAssocMap.Utils.flatten_assoc(@flatten_source)
   end
 
   test "splits by comma" do
@@ -61,14 +48,14 @@ defmodule PhpAssocMapTest do
       "'lvl_1_2'=>false"
     ]
 
-    assert PhpAssocMap.split_lines(@flatten_source) == expected
+    assert PhpAssocMap.Utils.split_lines(@flatten_source) == expected
   end
 
   test "splits entry by first key value" do
     value = "'lvl_1_2'=>['nested'=>'things']"
     expected = {"'lvl_1_2'", "['nested'=>'things']"}
 
-    assert PhpAssocMap.split_key_value(value) == expected
+    assert PhpAssocMap.Utils.split_key_value(value) == expected
   end
 
   test "parses associative array to map" do
@@ -81,6 +68,6 @@ defmodule PhpAssocMapTest do
       "lvl_1_2": false
     }
 
-    assert PhpAssocMap.to_map(@flatten_source) == expected
+    assert PhpAssocMap.to_tuple(@flatten_source) == expected
   end
 end
